+++ src/refactorings/extract-generic-type/extract-generic-type.interface.test.ts
@@ -1,10 +1,10 @@
-import { Code, Command } from "../../../editor/editor";
-import { InMemoryEditor } from "../../../editor/adapters/in-memory-editor";
-import { testEach } from "../../../tests-helpers";
+import { Code, Command } from "../../editor/editor";
+import { InMemoryEditor } from "../../editor/adapters/in-memory-editor";
+import { testEach } from "../../tests-helpers";
 
 import { extractGenericType } from "./extract-generic-type";
-import { ReplacementStrategy } from "../replacement-strategy";
-import { Position } from "../../../editor/position";
+import { ReplacementStrategy } from "../extract/replacement-strategy";
+import { Position } from "../../editor/position";
 
 describe("Extract Generic Type - Interface declaration", () => {
   testEach<{ code: Code; expected: Code }>(
