+++ src/refactorings/extract-generic-type/extract-generic-type.ts
@@ -1,13 +1,13 @@
-import { Editor, ErrorReason } from "../../../editor/editor";
-import { Selection } from "../../../editor/selection";
-import { Position } from "../../../editor/position";
-import * as t from "../../../ast";
+import { Editor, ErrorReason } from "../../editor/editor";
+import { Selection } from "../../editor/selection";
+import { Position } from "../../editor/position";
+import * as t from "../../ast";
 import {
   askReplacementStrategy,
   ReplacementStrategy
-} from "../replacement-strategy";
-import { renameSymbol } from "../../rename-symbol/rename-symbol";
-import { last } from "../../../array";
+} from "../extract/replacement-strategy";
+import { renameSymbol } from "../rename-symbol/rename-symbol";
+import { last } from "../../array";
 
 export { extractGenericType };
